Rename RegisterAndLogout to SingleBool
