+++ app/src/main/AndroidManifest.xml
@@ -2,6 +2,8 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     package="com.example.androidnewsui">
 
+    <uses-permission android:name="android.permission.INTERNET"/>
+
     <application
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
@@ -10,31 +12,34 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.AndroidNewsUI">
         <activity
-            android:name=".SelfinformationActivity"
+            android:name=".activity.SelfinformationActivity"
             android:exported="true" >
         </activity>
         <activity
-            android:name=".NewsaroundActivity"
+            android:name=".activity.NewsaroundActivity"
             android:exported="true">
+
+        </activity>
+        <activity
+            android:name=".activity.IndexActivity"
+            android:exported="true" >
+        </activity>
+        <activity
+            android:name=".activity.RegisterActivity"
+            android:exported="true" />
+        <activity
+            android:name=".activity.LoginActivity"
+            android:exported="true" />
+        <activity
+            android:name=".activity.MainActivity"
+            android:exported="true" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity
-            android:name=".IndexActivity"
-            android:exported="true" >
-        </activity>
-        <activity
-            android:name=".RegisterActivity"
-            android:exported="true" />
-        <activity
-            android:name=".LoginActivity"
-            android:exported="true" />
-        <activity
-            android:name=".MainActivity"
-            android:exported="true" />
+
     </application>
 
 </manifest>
