+++ app/src/main/java/com/example/androidnewsui/Adapter/MyAdapter.java
@@ -1,4 +1,4 @@
-package com.example.androidnewsui;
+package com.example.androidnewsui.Adapter;
 
 import android.content.Context;
 import android.view.LayoutInflater;
@@ -7,6 +7,9 @@ import android.view.ViewGroup;
 import android.widget.BaseAdapter;
 import android.widget.ImageView;
 import android.widget.TextView;
+
+import com.example.androidnewsui.NewsData;
+import com.example.androidnewsui.R;
 
 import java.util.LinkedList;
