+++ app/src/main/java/com/example/androidnewsui/activity/MainActivity.java
@@ -1,8 +1,10 @@
-package com.example.androidnewsui;
+package com.example.androidnewsui.activity;
+
+import android.os.Bundle;
 
 import androidx.appcompat.app.AppCompatActivity;
 
-import android.os.Bundle;
+import com.example.androidnewsui.R;
 
 public class MainActivity extends AppCompatActivity {
