+++ app/src/main/java/com/example/androidnewsui/activity/SelfinformationActivity.java
@@ -1,8 +1,10 @@
-package com.example.androidnewsui;
+package com.example.androidnewsui.activity;
 
 import androidx.appcompat.app.AppCompatActivity;
 
 import android.os.Bundle;
+
+import com.example.androidnewsui.R;
 
 public class SelfinformationActivity extends AppCompatActivity {
