+++ app/src/main/res/layout/activity_selfinformation.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".SelfinformationActivity">
+    tools:context=".activity.SelfinformationActivity">
 
     <LinearLayout
         android:layout_width="380dp"
